expanded User

+++ src/main/java/com/multitenant/multitenancy/meta/tenant/TenantRepo.java
@@ -4,7 +4,7 @@ import org.springframework.data.repository.CrudRepository;
 
 import java.util.stream.Stream;
 
-public interface TenantRepo extends CrudRepository<Tenant, Long> {
+public interface TenantRepo extends CrudRepository<Tenant, Integer> {
   Stream<Tenant> findAllByStateNot(TenantState state);
 
   void deleteBySchema(String schema);
